requirements.txt
monai
omegaconf
aim
pytorch-ignite
plotly
SimpleITK
nibabel
matplotlib!=3.6.1
scikit-image
scikit-learn
scipy
pandas
tqdm
torchvision
kornia
blobfile
mpi4py
fury
xvfbwrapper